+++ backend/Skarb.Api/Migrations/20260820131548_IgnoreDeletedSyncedAccounts.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 using Skarb.Api.Common.Persistence;
@@ -12,9 +13,11 @@ using Skarb.Api.Common.Persistence;
 namespace Skarb.Api.Migrations
 {
     [DbContext(typeof(SkarbDbContext))]
-    partial class SkarbDbContextModelSnapshot : ModelSnapshot
+    [Migration("20260820131548_IgnoreDeletedSyncedAccounts")]
+    partial class IgnoreDeletedSyncedAccounts
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder